% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vamplyr.R
\name{vamplyr_bela_lugosi}
\alias{vamplyr_bela_lugosi}
\title{Clearly the best Dracula: Bela Lugosi}
\usage{
vamplyr_bela_lugosi()
}
\value{
A line spoken by Bela Lugosi
}
\description{
Clearly the best Dracula: Bela Lugosi
}
\examples{
vamplyr_bela_lugosi()

}
