% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vamplyr.R
\name{vamplyr_gary_oldman}
\alias{vamplyr_gary_oldman}
\title{Clearly the best Dracula: Gary Oldman}
\usage{
vamplyr_gary_oldman()
}
\value{
A line spoken by Gary Oldman
}
\description{
Clearly the best Dracula: Gary Oldman
}
\examples{
vamplyr_gary_oldman()

}
